Создать iter1 (в процессе)
